phase 5 - sets refactoring, splitting logic apart and using ABCs
