+++ tests/test_battle_modifiers.py
@@ -3,11 +3,12 @@ import json
 from collections import defaultdict
 
 from fp import constants
-from fp.constants import BattleType
-from fp.modes import battle_mode
-from fp.data.pkmn_sets import (
-    TeamDatasets,
-    RandomBattleTeamDatasets,
+from fp.format_spec import FormatSpec
+from fp.modes.battle_factory import BattleFactoryMode
+from fp.modes.random_battle import RandomBattleMode
+from fp.modes.standard_battle import StandardBattleMode
+from fp.data.sets import (
+    BattleFactoryTeamDatasets,
     PredictedPokemonSet,
     PokemonSet,
     PokemonMoveset,
@@ -72,7 +73,7 @@ class TestRequestMessage(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.active = Pokemon("pikachu", 100)
         self.request_json = {
             "active": [
@@ -266,7 +267,7 @@ class TestSwitchOrDrag(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
         self.battle.user.active = Pokemon("pikachu", 100)
@@ -311,7 +312,7 @@ class TestSwitchOrDrag(unittest.TestCase):
 
     def test_does_not_add_pressure_to_impossible_abilities_gen3(self):
         self.battle.generation = "gen3"
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         split_msg = ["", "switch", "p2a: caterpie", "Caterpie, L100, M", "100/100"]
         switch_or_drag(self.battle, split_msg)
 
@@ -503,7 +504,7 @@ class TestSwitchOrDrag(unittest.TestCase):
 
     def test_increments_rest_turns_by_consequtive_sleeptalks(self):
         self.battle.generation = "gen3"
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         active = self.battle.opponent.active
         active.gen_3_consecutive_sleep_talks = 1
         active.rest_turns = 1
@@ -516,7 +517,7 @@ class TestSwitchOrDrag(unittest.TestCase):
 
     def test_decrements_sleep_turns_by_consequtive_sleeptalks(self):
         self.battle.generation = "gen3"
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         active = self.battle.opponent.active
         active.gen_3_consecutive_sleep_talks = 1
         active.sleep_turns = 1
@@ -551,7 +552,7 @@ class TestSwitchOrDrag(unittest.TestCase):
     def test_switch_does_not_reset_sleep_turns_to_0_in_gen4(self):
         self.battle.opponent.active.volatile_statuses.append(constants.TYPECHANGE)
         self.battle.generation = "gen4"
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         active = self.battle.opponent.active
         active.sleep_turns = 1
         active.status = constants.SLEEP
@@ -945,7 +946,7 @@ class TestHealOrDamage(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -1203,7 +1204,7 @@ class TestActivate(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -1339,7 +1340,7 @@ class TestPrepare(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -1364,7 +1365,7 @@ class TestClearAllBoosts(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -1407,7 +1408,7 @@ class TestMove(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -1416,15 +1417,15 @@ class TestMove(unittest.TestCase):
 
         self.battle.user.active = Pokemon("clefable", 100)
 
-        TeamDatasets.pkmn_sets = {}
-
     def test_infer_zoroark_from_move_not_possible_on_pkmn_battle_factory(self):
-        self.battle.mode = battle_mode(BattleType.BATTLE_FACTORY)
+        self.battle.mode = BattleFactoryMode()
         self.battle.generation = "gen9"
-        TeamDatasets.initialize(
-            "gen9battlefactory", ["zoroarkhisui", "gyarados"], "ru"
+        self.battle.mode.team_datasets = BattleFactoryTeamDatasets("ru")
+        self.battle.mode.team_datasets.initialize(
+            FormatSpec.from_format_string("gen9battlefactory"),
+            ["zoroarkhisui", "gyarados"],
         )  # gen9 RU should always have these pokemon
-        TeamDatasets.pkmn_sets["zoroarkhisui"] = [
+        self.battle.mode.team_datasets.pkmn_sets["zoroarkhisui"] = [
             PredictedPokemonSet(
                 pkmn_set=PokemonSet(
                     ability="illusion",
@@ -1478,9 +1479,11 @@ class TestMove(unittest.TestCase):
     def test_infer_zoroarkhisui_from_move_not_possible_on_pkmn_randbats_when_zoroark_unrevealed(
         self,
     ):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
 
         self.battle.opponent.active = Pokemon("gyarados", 79)
         self.battle.opponent.active.add_move("terablast")
@@ -1520,9 +1523,11 @@ class TestMove(unittest.TestCase):
     def test_infer_zoroark_regular_from_move_not_possible_on_pkmn_randbats_when_zoroark_unrevealed(
         self,
     ):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
 
         self.battle.opponent.active = Pokemon("gyarados", 79)
         self.battle.opponent.active.add_move("terablast")
@@ -1558,9 +1563,11 @@ class TestMove(unittest.TestCase):
     def test_does_not_infer_zoroark_if_move_can_be_on_active_pkmn(
         self,
     ):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
 
         self.battle.opponent.active = Pokemon("tornadustherian", 79)
         self.battle.opponent.active.add_move("terablast")
@@ -1581,9 +1588,11 @@ class TestMove(unittest.TestCase):
     def test_does_not_infer_zoroark_when_struggle(
         self,
     ):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
 
         self.battle.opponent.active = Pokemon("tornadustherian", 79)
         self.battle.opponent.active.add_move("terablast")
@@ -1602,10 +1611,12 @@ class TestMove(unittest.TestCase):
         self.assertEqual([], self.battle.opponent.reserve)
 
     def test_does_not_infer_from_struggle(self):
-        self.battle.mode = battle_mode(BattleType.BATTLE_FACTORY)
+        self.battle.mode = BattleFactoryMode()
         self.battle.generation = "gen9"
-        TeamDatasets.initialize(
-            "gen9battlefactory", ["zoroarkhisui", "gyarados"], "ru"
+        self.battle.mode.team_datasets = BattleFactoryTeamDatasets("ru")
+        self.battle.mode.team_datasets.initialize(
+            FormatSpec.from_format_string("gen9battlefactory"),
+            ["zoroarkhisui", "gyarados"],
         )  # gen9 RU should always have these pokemon
 
         self.battle.opponent.reserve = [Pokemon("zoroarkhisui", 100)]
@@ -2159,7 +2170,7 @@ class TestTrickRoom(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2196,7 +2207,7 @@ class TestWeather(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2454,7 +2465,7 @@ class TestSetBoost(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2500,7 +2511,7 @@ class TestBoostAndUnboost(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2592,7 +2603,7 @@ class TestStatus(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2667,7 +2678,7 @@ class TestCureStatus(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2724,7 +2735,7 @@ class TestStartFutureSight(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2752,7 +2763,7 @@ class TestSetItem(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -2835,7 +2846,7 @@ class TestStartVolatileStatus(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3075,7 +3086,7 @@ class TestEndVolatileStatus(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3233,7 +3244,7 @@ class TestUpdateAbility(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3369,7 +3380,7 @@ class TestSwapSideConditions(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3466,7 +3477,7 @@ class TestIllusionEnd(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3643,7 +3654,7 @@ class TestFail(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3671,7 +3682,7 @@ class TestFormChange(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3803,7 +3814,7 @@ class TestClearNegativeBoost(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3861,7 +3872,7 @@ class TestClearBoost(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3898,7 +3909,7 @@ class TestZPower(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3932,7 +3943,7 @@ class TestSideStart(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -3996,7 +4007,7 @@ class TestSingleTurn(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -4046,7 +4057,7 @@ class TestTransform(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -4278,7 +4289,7 @@ class TestCant(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -4341,7 +4352,7 @@ class TestUpkeep(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -4577,7 +4588,7 @@ class TestCheckSpeedRanges(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -5199,7 +5210,7 @@ class TestGuessChoiceScarf(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -5662,7 +5673,7 @@ class TestGuessChoiceScarf(unittest.TestCase):
         self.assertEqual("leftovers", self.battle.opponent.active.item)
 
     def test_uses_randombattle_spread_when_guessing_for_randombattle(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
 
         # opponent's speed should be 193 WITHOUT a choicescarf
         # HOWEVER, max-speed should still outspeed this value
@@ -5703,7 +5714,7 @@ class TestCheckHeavyDutyBoots(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -6148,7 +6159,7 @@ class TestRemoveItem(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -6190,7 +6201,7 @@ class TestImmune(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -6206,9 +6217,11 @@ class TestImmune(unittest.TestCase):
         self.battle.username = self.username
 
     def test_randbats_does_not_infer_zoroark_from_tera_immunity_on_judgment(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("enamorustherian", 83)
@@ -6230,9 +6243,11 @@ class TestImmune(unittest.TestCase):
         self.assertEqual(0, len(self.battle.opponent.reserve))
 
     def test_randbats_infer_zoroark_from_immunity_when_in_reserves(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
 
         self.battle.opponent.reserve = [Pokemon("zoroarkhisui", 80)]
         self.battle.opponent.reserve[0].add_move("nastyplot")
@@ -6271,9 +6286,11 @@ class TestImmune(unittest.TestCase):
         self.assertEqual({}, dict(self.battle.opponent.reserve[0].boosts))
 
     def test_randbats_infer_zoroarkhisui_from_immunity_when_not_in_reserves(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("gyarados", 100)
@@ -6306,9 +6323,11 @@ class TestImmune(unittest.TestCase):
         self.assertEqual({}, dict(self.battle.opponent.reserve[0].boosts))
 
     def test_randbats_infer_zoroark_from_immunity_when_not_in_reserves(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("gyarados", 100)
@@ -6341,9 +6360,9 @@ class TestImmune(unittest.TestCase):
         self.assertEqual({}, dict(self.battle.opponent.reserve[0].boosts))
 
     def test_gen4_does_not_infer_zoroark(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen4"
-        RandomBattleTeamDatasets.initialize("gen4")
+        self.battle.mode.datasets.initialize(FormatSpec.from_format_string("gen4"))
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("gyarados", 100)
@@ -6363,9 +6382,9 @@ class TestImmune(unittest.TestCase):
         self.assertEqual(0, len(self.battle.opponent.reserve))
 
     def test_gen5_does_not_infer_zoroark_hisui(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen5"
-        RandomBattleTeamDatasets.initialize("gen5")
+        self.battle.mode.datasets.initialize(FormatSpec.from_format_string("gen5"))
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("gyarados", 100)
@@ -6385,9 +6404,11 @@ class TestImmune(unittest.TestCase):
         self.assertEqual(0, len(self.battle.opponent.reserve))
 
     def test_does_not_infer_zoroark_if_pkmn_terastallized_to_gain_immunity(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("gyarados", 100)
@@ -6406,9 +6427,11 @@ class TestImmune(unittest.TestCase):
         self.assertEqual(0, len(self.battle.opponent.reserve))
 
     def test_does_not_infer_zoroark_if_pkmn_naturally_immune(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("urshifu", 100)
@@ -6425,9 +6448,11 @@ class TestImmune(unittest.TestCase):
         self.assertEqual(0, len(self.battle.opponent.reserve))
 
     def test_does_not_infer_zoroark_if_futuresight_ending(self):
-        self.battle.mode = battle_mode(BattleType.RANDOM_BATTLE)
+        self.battle.mode = RandomBattleMode()
         self.battle.generation = "gen9"
-        RandomBattleTeamDatasets.initialize("gen9randombattle")
+        self.battle.mode.datasets.initialize(
+            FormatSpec.from_format_string("gen9randombattle")
+        )
         self.battle.opponent.reserve = []
 
         self.battle.opponent.active = Pokemon("Urshifu", 100)
@@ -6445,10 +6470,12 @@ class TestImmune(unittest.TestCase):
         self.assertEqual(0, len(self.battle.opponent.reserve))
 
     def test_infers_zoroark_from_immunity_that_pkmn_does_not_have(self):
-        self.battle.mode = battle_mode(BattleType.BATTLE_FACTORY)
+        self.battle.mode = BattleFactoryMode()
         self.battle.generation = "gen9"
-        TeamDatasets.initialize(
-            "gen9battlefactory", ["zoroarkhisui", "gyarados"], "ru"
+        self.battle.mode.team_datasets = BattleFactoryTeamDatasets("ru")
+        self.battle.mode.team_datasets.initialize(
+            FormatSpec.from_format_string("gen9battlefactory"),
+            ["zoroarkhisui", "gyarados"],
         )  # gen9 RU should always have these pokemon
 
         self.battle.opponent.reserve = [Pokemon("zoroarkhisui", 100)]
@@ -6486,10 +6513,12 @@ class TestImmune(unittest.TestCase):
         self.assertEqual({}, dict(self.battle.opponent.reserve[0].boosts))
 
     def test_does_not_infer_zoroark_when_tera_type_renders_it_immune(self):
-        self.battle.mode = battle_mode(BattleType.BATTLE_FACTORY)
+        self.battle.mode = BattleFactoryMode()
         self.battle.generation = "gen9"
-        TeamDatasets.initialize(
-            "gen9battlefactory", ["zoroarkhisui", "gyarados"], "ru"
+        self.battle.mode.team_datasets = BattleFactoryTeamDatasets("ru")
+        self.battle.mode.team_datasets.initialize(
+            FormatSpec.from_format_string("gen9battlefactory"),
+            ["zoroarkhisui", "gyarados"],
         )  # gen9 RU should always have these pokemon
 
         self.battle.opponent.reserve = [Pokemon("zoroarkhisui", 100)]
@@ -6511,10 +6540,12 @@ class TestImmune(unittest.TestCase):
         self.assertEqual("zoroarkhisui", self.battle.opponent.reserve[0].name)
 
     def test_does_not_infer_zoroark_when_pkmn_is_actually_immune(self):
-        self.battle.mode = battle_mode(BattleType.BATTLE_FACTORY)
+        self.battle.mode = BattleFactoryMode()
         self.battle.generation = "gen9"
-        TeamDatasets.initialize(
-            "gen9battlefactory", ["zoroarkhisui", "maushold"], "ru"
+        self.battle.mode.team_datasets = BattleFactoryTeamDatasets("ru")
+        self.battle.mode.team_datasets.initialize(
+            FormatSpec.from_format_string("gen9battlefactory"),
+            ["zoroarkhisui", "maushold"],
         )  # gen9 RU should always have these pokemon
 
         self.battle.opponent.reserve = [Pokemon("zoroarkhisui", 100)]
@@ -6533,10 +6564,12 @@ class TestImmune(unittest.TestCase):
         self.assertEqual("zoroarkhisui", self.battle.opponent.reserve[0].name)
 
     def test_does_not_infer_zoroark_when_the_zoroark_is_not_immune(self):
-        self.battle.mode = battle_mode(BattleType.BATTLE_FACTORY)
+        self.battle.mode = BattleFactoryMode()
         self.battle.generation = "gen9"
-        TeamDatasets.initialize(
-            "gen9battlefactory", ["zoroarkhisui", "salamence"], "ru"
+        self.battle.mode.team_datasets = BattleFactoryTeamDatasets("ru")
+        self.battle.mode.team_datasets.initialize(
+            FormatSpec.from_format_string("gen9battlefactory"),
+            ["zoroarkhisui", "salamence"],
         )  # gen9 RU should always have these pokemon
 
         self.battle.opponent.reserve = [Pokemon("zoroarkhisui", 100)]
@@ -6555,10 +6588,12 @@ class TestImmune(unittest.TestCase):
         self.assertEqual("zoroarkhisui", self.battle.opponent.reserve[0].name)
 
     def test_does_not_infer_zoroark_when_ability_renders_immune(self):
-        self.battle.mode = battle_mode(BattleType.BATTLE_FACTORY)
+        self.battle.mode = BattleFactoryMode()
         self.battle.generation = "gen9"
-        TeamDatasets.initialize(
-            "gen9battlefactory", ["zoroarkhisui", "rotomheat"], "ru"
+        self.battle.mode.team_datasets = BattleFactoryTeamDatasets("ru")
+        self.battle.mode.team_datasets.initialize(
+            FormatSpec.from_format_string("gen9battlefactory"),
+            ["zoroarkhisui", "rotomheat"],
         )  # gen9 RU should always have these pokemon
 
         self.battle.opponent.reserve = [Pokemon("zoroarkhisui", 100)]
@@ -6598,7 +6633,7 @@ class TestInactive(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -6651,7 +6686,7 @@ class TestInactiveOff(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
         self.battle.user.name = "p1"
         self.battle.opponent.name = "p2"
 
@@ -6698,7 +6733,7 @@ class TestGetDamageDealt(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
 
         self.battle.user.name = "p1"
         self.battle.user.active = Pokemon("Caterpie", 100)
@@ -7060,7 +7095,7 @@ class TestNoInit(unittest.TestCase):
     def setUp(self):
         self.battle = Battle(None)
         self.battle.generation = "gen9"
-        self.battle.mode = battle_mode(BattleType.STANDARD_BATTLE)
+        self.battle.mode = StandardBattleMode()
 
         self.battle.user.name = "p1"
         self.battle.user.active = Pokemon("Caterpie", 100)
